[#23] Fix off-by-one in error offsets
Problem: our errors currently pretty-print nicely, except that both
the line and column numbers and the pointer are off-by-one, which
makes the error reporting less useful. The reason for this is that
the initial source position for the parser state is initialised with
zero for both line and column number, whereas it should be, and is
in the reference implementation, set to 1.

Solution: changed the initial positions to 1.
Also changed all Bit references to the one from YatimaStdLib.